move docsrating to components
