+++ src/components/DocsRating.js
