+++ src/theme/DocItem/Footer/index.js
@@ -7,7 +7,7 @@ import EditThisPage from '@theme/EditThisPage';
 import TagsListInline from '@theme/TagsListInline';
 
 import styles from './styles.module.css';
-import DocsRating from '../../../../scripts/docs-rating';
+import DocsRating from '../../../../src/components/DocsRating';
 
 function TagsRow(props) {
   return (
